Export module interface
The systx is a transaction management system. This patch adds an
interface for modules to extend systx with end-user functionality.

The function systx_register_module() registers a new module. The call
returns the number assigned to the registered module on success, or
an errno code otherwise. The existing components will still have to
be converted to the new interface.

The function systx_inject_event() allows modules to put events into
the transaction log during an executed transaction. Systx uses these
event to apply or undo operations when the transaction commits or
aborts.

A module shall call systx_resolve_conflict() if it detects a conflict
between transactions. Systx will try to resolve the conflict, possibly
by restarting one of the involved transactions.

A module shall call systx_resolve_error() if it detects an internal
run-time error. Systx will try to repair the error, restart, or
ultimatively abort the transaction.