tm: Add |struct tm_frame_map| data structure
The frame map is a data structure that maps memory addresses to their
corresponding frames. It's implemented as sparse tree, similar to the
processor's page table. This layout makes it more memory efficient than
the linear frame table.

Lookups in the frame map have a higher processor overhead than lookups
in the linear frame table, but the addresses of returned frame don't
change. This allows for caching frames in a transaction-local working
set.