tm: Add write-through mode for memory pages
By default, load and store operations work on the private buffers
of memory pages, as represented by |struct tm_page|. The changes are
writen back to memory when applying the commit.

This patch adds write-though mode. Loads and stores operate directly
on memory, and changes are reverted when rolling back a transaction.

This feature is useful for privatizing memory regions, or avoiding
commit overhead on memory regions with low contention.