libc: Add support for 'errno' in transactions
'errno' is a thread-local variable that stores the error code returned
by the last failed system call or function of the C standard library.

This patch adds support for saving the value of errno, and restoring
it when a transaction is aborted. Because 'errno' is a thread-local
variable, no concurrency control is required.