Rename picotm_resolve_error() to picotm_recover_from_errno()
This patch renames picotm_resolve_error(). The new name reflects the
nature of the resolution, which is a recovery, and explicitly mentions
errno codes. Similar functions can be added for other systems of error
codes.

The patch also fixes several occations where an incorrect or negative
errno code was passed to the function.