Introduce module-wide apply and undo functions
This patch adds module-wide apply and undo functions. Modules that that
don't perform destructive operations can register module-wide apply and
undo functions to perform a single apply or undo *before* all events.

The new module call-back functions are useful for modules such as TM,
which before had to append pseudo-events to the log to get a call for
a single apply or undo.