Call module apply and undo callbacks with single events only
This patch removes the array-length argument from the module's apply
and undo call-back functions. The callbacks are now invoked for each
event in the log specifically.

This change simplifies the event handling in the picotm framework and
modules. Modules that want to merge consecutive events, can do so by
implementing the module-wide apply and undo call-back functions.