Move most of picotm_error_*() functions into C source file
This patch moves most of the picotm_error_*() functions from the header
file into the implementation. These fucntions are rarely used and not
critical to performance.

The only execption is picom_error_is_set(), which remains in the header
file. The function is the standard test for errors and is likely to be
called in hot code paths. Its implementation is just a single line, so
inlining the function is probably best in any case.