libc: Maintain list of active ofd_tx structures for each transaction
When operating on ofd_tx instances, the code for |struct fildes_tx| used
to walk over the whole array of |struct ofd_tx| instances, filtering out
those instance that are used by the transaction. This strategy is wasteful
in terms of memory consumption and computing time.

This patch adds a list of active instances of |struct ofd_tx| to
|struct fildes_tx|. When an instance of |struct ofd_tx| is first acquired
by a transaction, it's added to the list. All later operation of
|struct fildes_tx| use this list to walk over the list of active instances
of |struct ofd_tx|.