libc: Fail with EBADF for using file from different file descriptors
Using the same file with different file descriptors concurrently will
now abort the transaction with an EBADF errno code. This limitation
guarantees consistency of the involved open file descriptions.

There's an exception for file descriptors created with in a transaction
(e.g., by open() or dup()). These are allowed to refer to existing file
buffer, because we know which open file description is involved.

This patch adapts test cases fdio-8 and fdio-15 to work with the new
semantics.