tests: Move test-tool infrastructure to libraries
This patch moves the common infrastructure of all test cases to
two libraries. The library in 'tests/libtests' implements the common
safe_* helpers, thread handling, and test loops. The library in
'tests/libtap' implements the TAP output.

The TAP library is self-contained and independent from the rest of
the package. It can be reused for other software.