Garbage-collect unused symbols when linking binaries
This patch adds support for compiler and linker flags that can reduce
the size of the produced binary files significantly. For statically
linked binaries, only the actually used functions and data is included
by the linker. The unlinked static library may be much larger though.
There's only a small effect on shared objects.