lib: Store jump buffer on thread-local stack
The jump buffer for returning to the beginning of the transaction
was stored in the thread-local transaction state. Initializaton of
this state can fail and the transaction would roll back. Since there
is not yet a valid jump buffer, the roll-back would fail and move
into an undefined state.

This patch changes picotm's interface and implementation such that
the jump buffer is stored on the thread's stack before entering a
transaction and it's initialized before the transaction state. If
initialization of the transaction state fails, the transaction can
will now enter recovery mode directly.