libc: Decode `mode_t` as type `int` in va_arg()
The type of `mode_t` might have a conversion rank lower than `int`. A
parameter of `mode_t` will be promoted to `int`. Decoding variable-
argument parameters as `mode_t` then results in undefined behaviour. This
is the case on at least FreeBSD systems.

This patch changes the call to `va_arg()` to decode `mode_t` as `int`
when passed to `open_tx()` and `open_tm()`.