doc: Define __PICOTM_DOXYGEN when pre-processing documentation
C header files can contain declarations that are covered by an #if
pre-processor conditional. These are often dependend on the system
and can get lost when building the documentation. The defined pre-
processor token __PICOTM_DOXYGEN can be par tof the condition, such
that the declaration shows up in the documenation, as illustrated
below.

\#if condition || defined(__PICOTM_DOXYGEN)
/** documentation */
void system_dependent_func_decl(void)
\#endif

The compiler only sees this declaration if the condition is true. A
doxygen run always sees it, as __PICOTM_DOXYGEN is defined.