core: Convert |struct global_state| to ref-counted object
This patch adds reference counting to |struct global_state|. Each thread
holds a reference to the global state, which it releases as part of its
own cleanup. The global state will therefore be around no longer than
its users.

As a side effect, the use of `atexit()` has been removed. This benefits
the application, as the number of atexit slots is limited.