core: Add PICOTM_MODE_RESTART for retries from recovery mode
Restarting from recvery mode was done with mode PICOTM_MODE_RETRY.
This mode implies a roll back of the transaction log, which is not
the case for recovery. The new mode PICOTM_MODE_RESTART is a retry
from recovery; without an implicit rollback.

As a consequence, this patch also removes 'do_rollback' from the
arguments of restart_tx(). This is information is now given by the
mode directly.