Mass cleanup of include statements
Building picotm, its modules and tests requiries the use of the internal
headers that come with the package. Already installed header files of
previous releases can interfere with this.

This patch cleans up the include statements of all internal headers.
These headers are now included with quotes, which allow for additional
search paths in the C preprocessor. The search paths are specified with
the -iquote option.