libc/fildes: Fail in compare if file descriptor is non-equal in file tables
On Unix we cannot distinguish between files, because open file
descriptions do not have a unique id. Instead the composed file id
consists of a file buffer and file descriptors.

File descriptors to the same file buffer can either share their file
structure, or not. To make picotm work reliably, we have to expect that
file descriptors to the same file buffer always share their file
structure. Therefore fail in look-up if the key id has a file descriptor
that is different from the stored one.

Signed-off-by: Thomas Zimmermann <contact@tzimmermann.org>