libc/fildes: Compare files for equality, not order
Ordering open file descriptions is a fuzzy concept at best. It is not
implemented and support from the operating system is questionable. So
let's not pretend we can do that.

Rename the file id _cmp functions to _cmp_eq and return true or false,
depending on equality of the given files. Update callers accordingly.

Signed-off-by: Thomas Zimmermann <tdz@picotm.org>